Refactor sstate files to be clearer; remove unnecessary build tasks
Change the naming of the files and the YAML content to be more consistent and
clearer on what it's intended to do. Also remove old build "test" tasks that
were used to validate that the sstate deployment worked in the first place,
since the actual build pipelines in meta-python and poky can already do this.

Signed-off-by: Trevor Gamblin <tvgamblin@gmail.com>