Add whenExpression for meta-python-log, Result for meta-python-build
Turns out that the pipeline was breaking if COMMIT_LOG was empty, which would
occur if there was no identified delta between master and master-next. I've
added a rudimentary Result and whenExpression to the process to catch this. In
the future, we could potentially use the same idea to break the actual checking
of the git log out into a separate step/task, and make the COMMIT_LOG contents
be a result/input from elsewhere.

Signed-off-by: Trevor Gamblin <tvgamblin@gmail.com>